Renamed component files to index
